Allow to specify mink driver in browser configuration.
* Add option to specify the mink driver as part of the browser configuration.
* For backward compatibility defaults to Selenium2Driver.
* Add Goutte and SahiDriver support.
* Allows to register additional mink drivers.

For example, to configure a test to run with the Goutte driver, use

    public static $browsers = [[
        'browserName' => 'Goutte',
        'driver' => 'goutte'
    ]];

There also is a new configuration option 'driverOptions' which can be used
by driver factories when instantiating the mink driver.
It defaults to null, places no restriction on the data structure and
is currently not used by any of the driver factories.

This PR is realated to issue #12.