Allow to specify Mink driver in browser configuration
* add option for specifying Mink driver as part of the browser configuration
* set driver to 'selenium2' to preserve backwards compatibility
* add driver factories for Selenium2, Zombie, Goutte and Sahi drivers (only basic options supported)
* allows to register additional Mink drivers

For example, to configure a test to run with the Goutte driver, use

public static $browsers = array(
	array(
		'driver' => 'goutte',
    )
);

There also is a new configuration option 'driverOptions' which can be used by driver factories when instantiating the Mink driver.
It defaults to an empty arrray, places no restriction on the data structure and is currently not used by any of the driver factories.